pyproject.toml
[project]
name = "leasure"
version = "0.1.0"
description = "Music downloader and MP3 player sync for HIFI WALKER H2"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.7",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.7",
    "spotdl>=4.2",
    "streamrip>=2.0",
    "spotipy>=2.24",
    "ytmusicapi>=1.9",
    "bandcamp-downloader",
    "Pillow>=12.1.1",
    "mutagen>=1.47",
    "httpx>=0.28",
]

[project.scripts]
leasure = "app:main"
